+++ example/src/main/java/org/pfruender/mockito/example/SampleDependentServiceOne.java
@@ -13,8 +13,9 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.hypoport.mockito.example;
+package org.pfruender.mockito.example;
 
-public class ResultOfDependentServiceTwo {
+public interface SampleDependentServiceOne {
 
+  ResultOfDependentServiceOne getResult(ServiceInputParameter param1, ServiceInputParameter param2);
 }
