+++ example/src/main/java/org/pfruender/mockito/example/SampleDependentServiceTwo.java
@@ -13,20 +13,9 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.hypoport.mockito.example;
+package org.pfruender.mockito.example;
 
-import javax.inject.Inject;
+public interface SampleDependentServiceTwo {
 
-public class SampleOrchestratingService {
-
-  @Inject
-  SampleDependentServiceOne dependentServiceOne;
-
-  @Inject
-  SampleDependentServiceTwo dependentServiceTwo;
-
-  public ResultOfDependentServiceTwo doService(ServiceInputParameter param1, ServiceInputParameter param2) {
-    ResultOfDependentServiceOne result1 = dependentServiceOne.getResult(param1, param2);
-    return dependentServiceTwo.getResult(result1);
-  }
+  ResultOfDependentServiceTwo getResult(ResultOfDependentServiceOne result1);
 }
