+++ example/src/main/java/org/pfruender/mockito/example/SampleOrchestratingService.java
@@ -13,11 +13,20 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.hypoport.mockito.provider;
+package org.pfruender.mockito.example;
 
-public interface IProvided {
+import javax.inject.Inject;
 
-  String getSomething();
+public class SampleOrchestratingService {
 
-  Object getObject();
+  @Inject
+  SampleDependentServiceOne dependentServiceOne;
+
+  @Inject
+  SampleDependentServiceTwo dependentServiceTwo;
+
+  public ResultOfDependentServiceTwo doService(ServiceInputParameter param1, ServiceInputParameter param2) {
+    ResultOfDependentServiceOne result1 = dependentServiceOne.getResult(param1, param2);
+    return dependentServiceTwo.getResult(result1);
+  }
 }
