+++ example/src/main/java/org/pfruender/mockito/example/ServiceInputParameter.java
@@ -13,9 +13,8 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.hypoport.mockito.example;
+package org.pfruender.mockito.example;
 
-public interface SampleDependentServiceTwo {
+public class ServiceInputParameter {
 
-  ResultOfDependentServiceTwo getResult(ResultOfDependentServiceOne result1);
 }
