+++ src/main/java/org/pfruender/mockito/MockInjectorConfigurator.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.hypoport.mockito;
+package org.pfruender.mockito;
 
 import java.lang.annotation.Annotation;
 import java.util.Collection;
