+++ src/test/java/org/pfruender/mockito/injection/Injected.java
@@ -13,9 +13,8 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.hypoport.mockito.example;
+package org.pfruender.mockito.injection;
 
-public interface SampleDependentServiceOne {
+public interface Injected {
 
-  ResultOfDependentServiceOne getResult(ServiceInputParameter param1, ServiceInputParameter param2);
 }
