+++ src/test/java/org/pfruender/mockito/injection/MockInjectorTest.java
@@ -13,9 +13,9 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.hypoport.mockito.injection;
+package org.pfruender.mockito.injection;
 
-import org.hypoport.mockito.MockInjector;
+import org.pfruender.mockito.MockInjector;
 import org.testng.annotations.Test;
 
 import java.lang.reflect.Field;
