+++ src/test/java/org/pfruender/mockito/injection/MyClass.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.hypoport.mockito.injection;
+package org.pfruender.mockito.injection;
 
 import javax.annotation.Resource;
 import javax.inject.Inject;
